+++ src/base/b-virtual-scroll-new/const.ts
@@ -177,12 +177,6 @@ export const defaultShouldProps = <const>{
 		return isLastRequestEmpty();
 	},
 
-	/** {@link bVirtualScroll.shouldPerformDataRequest} */
-	shouldPerformDataRequest: (state: VirtualScrollState, _ctx: bVirtualScrollNew): boolean => {
-		const isLastRequestNotEmpty = () => state.lastLoadedData.length > 0;
-		return isLastRequestNotEmpty();
-	},
-
 	/** {@link bVirtualScroll.shouldPerformDataRender} */
 	shouldPerformDataRender: (state: VirtualScrollState, _ctx: bVirtualScrollNew): boolean =>
 		state.isInitialRender || state.remainingItems === 0